Attempt at fixing uncaught crash
Attempt at fixing once and for all any crash related to reading/writing memory when it's not accessible.

The python stub should still be used to not perform the action, but if the action is performed anyway, it should just panic and not crash anymore. However it could crash later if the read/write was necessary for the game to run normally.